Implementation of Mont Carlo algorithm related to finance on GPU
Mont Carlo algorithm has verity of application especially in finance
which could predict the value of stocks. This algorithm due to its huge
amount of computation, demands on devices could compute a large amount
of operation simultaneously to meet the time and energy-efficiency
constraints. In this report we assert a GPU-based implementation of Mont
Carlo algorithm related to financial issues. GPUs could pave the way for
us to leverage most sources of Parallelism in Mont Carlo.